Refactored the testsuite + added the proper KeycloakRBACAuthorizer grants refresh test
